+++ libsparse/output_file.c
@@ -61,30 +61,30 @@
 
 struct output_file_ops {
 	int (*open)(struct output_file *, int fd);
-	int (*skip)(struct output_file *, int64_t);
-	int (*pad)(struct output_file *, int64_t);
-	int (*write)(struct output_file *, void *, int);
+	int (*skip)(struct output_file *, off_t);
+	int (*pad)(struct output_file *, ssize_t);
+	int (*write)(struct output_file *, void *, ssize_t);
 	void (*close)(struct output_file *);
 };
 
 struct sparse_file_ops {
-	int (*write_data_chunk)(struct output_file *out, unsigned int len,
+	int (*write_data_chunk)(struct output_file *out, ssize_t len,
 			void *data);
-	int (*write_fill_chunk)(struct output_file *out, unsigned int len,
+	int (*write_fill_chunk)(struct output_file *out, ssize_t len,
 			uint32_t fill_val);
-	int (*write_skip_chunk)(struct output_file *out, int64_t len);
+	int (*write_skip_chunk)(struct output_file *out, ssize_t len);
 	int (*write_end_chunk)(struct output_file *out);
 };
 
 struct output_file {
-	int64_t cur_out_ptr;
+	off_t cur_out_ptr;
 	unsigned int chunk_cnt;
 	uint32_t crc32;
 	struct output_file_ops *ops;
 	struct sparse_file_ops *sparse_ops;
 	int use_crc;
 	unsigned int block_size;
-	int64_t len;
+	ssize_t len;
 	char *zero_buf;
 	uint32_t *fill_buf;
 	char *buf;
@@ -109,7 +109,7 @@ struct output_file_normal {
 struct output_file_callback {
 	struct output_file out;
 	void *priv;
-	int (*write)(void *priv, const void *buf, int len);
+	int (*write)(void *priv, const void *buf, ssize_t len);
 };
 
 #define to_output_file_callback(_o) \
@@ -123,7 +123,7 @@ static int file_open(struct output_file *out, int fd)
 	return 0;
 }
 
-static int file_skip(struct output_file *out, int64_t cnt)
+static int file_skip(struct output_file *out, off_t cnt)
 {
 	off64_t ret;
 	struct output_file_normal *outn = to_output_file_normal(out);
@@ -136,7 +136,7 @@ static int file_skip(struct output_file *out, int64_t cnt)
 	return 0;
 }
 
-static int file_pad(struct output_file *out, int64_t len)
+static int file_pad(struct output_file *out, ssize_t len)
 {
 	int ret;
 	struct output_file_normal *outn = to_output_file_normal(out);
@@ -149,7 +149,7 @@ static int file_pad(struct output_file *out, int64_t len)
 	return 0;
 }
 
-static int file_write(struct output_file *out, void *data, int len)
+static int file_write(struct output_file *out, void *data, ssize_t len)
 {
 	int ret;
 	struct output_file_normal *outn = to_output_file_normal(out);
@@ -195,7 +195,7 @@ static int gz_file_open(struct output_file *out, int fd)
 }
 
 
-static int gz_file_skip(struct output_file *out, int64_t cnt)
+static int gz_file_skip(struct output_file *out, off_t cnt)
 {
 	off64_t ret;
 	struct output_file_gz *outgz = to_output_file_gz(out);
@@ -208,7 +208,7 @@ static int gz_file_skip(struct output_file *out, int64_t cnt)
 	return 0;
 }
 
-static int gz_file_pad(struct output_file *out, int64_t len)
+static int gz_file_pad(struct output_file *out, ssize_t len)
 {
 	off64_t ret;
 	struct output_file_gz *outgz = to_output_file_gz(out);
@@ -232,7 +232,7 @@ static int gz_file_pad(struct output_file *out, int64_t len)
 	return 0;
 }
 
-static int gz_file_write(struct output_file *out, void *data, int len)
+static int gz_file_write(struct output_file *out, void *data, ssize_t len)
 {
 	int ret;
 	struct output_file_gz *outgz = to_output_file_gz(out);
@@ -270,7 +270,7 @@ static int callback_file_open(struct output_file *out __unused, int fd __unused)
 	return 0;
 }
 
-static int callback_file_skip(struct output_file *out, int64_t off)
+static int callback_file_skip(struct output_file *out, off_t off)
 {
 	struct output_file_callback *outc = to_output_file_callback(out);
 	int to_write;
@@ -288,12 +288,12 @@ static int callback_file_skip(struct output_file *out, int64_t off)
 	return 0;
 }
 
-static int callback_file_pad(struct output_file *out __unused, int64_t len __unused)
+static int callback_file_pad(struct output_file *out __unused, ssize_t len __unused)
 {
 	return -1;
 }
 
-static int callback_file_write(struct output_file *out, void *data, int len)
+static int callback_file_write(struct output_file *out, void *data, ssize_t len)
 {
 	struct output_file_callback *outc = to_output_file_callback(out);
 
@@ -337,7 +337,7 @@ int read_all(int fd, void *buf, size_t len)
 	return 0;
 }
 
-static int write_sparse_skip_chunk(struct output_file *out, int64_t skip_len)
+static int write_sparse_skip_chunk(struct output_file *out, off_t skip_len)
 {
 	chunk_header_t chunk_header;
 	int ret;
@@ -363,7 +363,7 @@ static int write_sparse_skip_chunk(struct output_file *out, int64_t skip_len)
 	return 0;
 }
 
-static int write_sparse_fill_chunk(struct output_file *out, unsigned int len,
+static int write_sparse_fill_chunk(struct output_file *out, ssize_t len,
 		uint32_t fill_val)
 {
 	chunk_header_t chunk_header;
@@ -398,7 +398,7 @@ static int write_sparse_fill_chunk(struct output_file *out, unsigned int len,
 	return 0;
 }
 
-static int write_sparse_data_chunk(struct output_file *out, unsigned int len,
+static int write_sparse_data_chunk(struct output_file *out, ssize_t len,
 		void *data)
 {
 	chunk_header_t chunk_header;
@@ -472,7 +472,7 @@ static struct sparse_file_ops sparse_file_ops = {
 		.write_end_chunk = write_sparse_end_chunk,
 };
 
-static int write_normal_data_chunk(struct output_file *out, unsigned int len,
+static int write_normal_data_chunk(struct output_file *out, ssize_t len,
 		void *data)
 {
 	int ret;
@@ -490,7 +490,7 @@ static int write_normal_data_chunk(struct output_file *out, unsigned int len,
 	return ret;
 }
 
-static int write_normal_fill_chunk(struct output_file *out, unsigned int len,
+static int write_normal_fill_chunk(struct output_file *out, ssize_t len,
 		uint32_t fill_val)
 {
 	int ret;
@@ -515,7 +515,7 @@ static int write_normal_fill_chunk(struct output_file *out, unsigned int len,
 	return 0;
 }
 
-static int write_normal_skip_chunk(struct output_file *out, int64_t len)
+static int write_normal_skip_chunk(struct output_file *out, ssize_t len)
 {
 	return out->ops->skip(out, len);
 }
@@ -539,7 +539,7 @@ void output_file_close(struct output_file *out)
 }
 
 static int output_file_init(struct output_file *out, int block_size,
-		int64_t len, bool sparse, int chunks, bool crc)
+		ssize_t len, bool sparse, int chunks, bool crc)
 {
 	int ret;
 
@@ -627,8 +627,8 @@ static struct output_file *output_file_new_normal(void)
 	return &outn->out;
 }
 
-struct output_file *output_file_open_callback(int (*write)(void *, const void *, int),
-		void *priv, unsigned int block_size, int64_t len,
+struct output_file *output_file_open_callback(int (*write)(void *, const void *, ssize_t),
+		void *priv, unsigned int block_size, ssize_t len,
 		int gz __unused, int sparse, int chunks, int crc)
 {
 	int ret;
@@ -653,7 +653,7 @@ struct output_file *output_file_open_callback(int (*write)(void *, const void *,
 	return &outc->out;
 }
 
-struct output_file *output_file_open_fd(int fd, unsigned int block_size, int64_t len,
+struct output_file *output_file_open_fd(int fd, unsigned int block_size, ssize_t len,
 		int gz, int sparse, int chunks, int crc)
 {
 	int ret;
@@ -680,25 +680,25 @@ struct output_file *output_file_open_fd(int fd, unsigned int block_size, int64_t
 }
 
 /* Write a contiguous region of data blocks from a memory buffer */
-int write_data_chunk(struct output_file *out, unsigned int len, void *data)
+int write_data_chunk(struct output_file *out, ssize_t len, void *data)
 {
 	return out->sparse_ops->write_data_chunk(out, len, data);
 }
 
 /* Write a contiguous region of data blocks with a fill value */
-int write_fill_chunk(struct output_file *out, unsigned int len,
+int write_fill_chunk(struct output_file *out, ssize_t len,
 		uint32_t fill_val)
 {
 	return out->sparse_ops->write_fill_chunk(out, len, fill_val);
 }
 
-int write_fd_chunk(struct output_file *out, unsigned int len,
-		int fd, int64_t offset)
+int write_fd_chunk(struct output_file *out, ssize_t len,
+		int fd, off_t offset)
 {
 	int ret;
-	int64_t aligned_offset;
-	int aligned_diff;
-	int buffer_size;
+	off64_t aligned_offset;
+	off_t aligned_diff;
+	ssize_t buffer_size;
 	char *ptr;
 
 	aligned_offset = offset & ~(4096 - 1);
@@ -743,8 +743,8 @@ int write_fd_chunk(struct output_file *out, unsigned int len,
 }
 
 /* Write a contiguous region of data blocks from a file */
-int write_file_chunk(struct output_file *out, unsigned int len,
-		const char *file, int64_t offset)
+int write_file_chunk(struct output_file *out, ssize_t len,
+		const char *file, off_t offset)
 {
 	int ret;
 
@@ -760,7 +760,7 @@ int write_file_chunk(struct output_file *out, unsigned int len,
 	return ret;
 }
 
-int write_skip_chunk(struct output_file *out, int64_t len)
+int write_skip_chunk(struct output_file *out, ssize_t len)
 {
 	return out->sparse_ops->write_skip_chunk(out, len);
 }
